Collect Rpc response types
