+++ client/src/lib.rs
@@ -8,4 +8,5 @@ pub mod perf_utils;
 pub mod rpc_client;
 pub mod rpc_client_request;
 pub mod rpc_request;
+pub mod rpc_response;
 pub mod thin_client;
